:tada: make v14 work with commonjs or module loader
