+++ .gitignore
@@ -3,3 +3,5 @@ v10.x/layer
 v10.x/test/lambda.zip
 v12.x/layer
 v12.x/test/lambda.zip
+v14.x/layer
+v14.x/test/**/lambda.zip
