+++ v14.x/test/commonjs/data.json
@@ -1,0 +1,3 @@
+{
+  "message": "complete"
+}
