+++ v14.x/test/commonjs/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "test",
+  "version": "1.0.0",
+  "dependencies": {
+    "aws4": "^1.8.0"
+  }
+}
