+++ v14.x/test/module/data.json
@@ -1,0 +1,3 @@
+{
+  "message": "complete"
+}
