+++ v14.x/test/module/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "test",
+  "version": "1.0.0",
+  "type": "module",
+  "dependencies": {
+    "aws4": "^1.8.0"
+  }
+}
