--- v14.x/test/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "test",
-  "version": "1.0.0",
-  "type": "module",
-  "dependencies": {
-    "aws4": "^1.8.0"
-  }
-}
